Friends timeline button working. Initial and refresh uncoupled.
Originally, refreshMessages was called even on the initial load. In anticipation of displaying more than just the friends timeline, I
made a getFriends function that is to be called on initial load and
when the Friends button is clicked. The gettweets function now takes
a parameter that determines which set of statuses is retrieved.