Deleted tweets are immediately invisible. New tweets don't always reset the display.
If a new tweet is added from 'friends' or 'mine', the display is refreshed with that tweet. If a new tweet is added from 'replies' or 'favorites', the display is reset to 'friends'. A deleted tweet is not just sent a destroy message at Twitter, it's display here is set to none.